+++ src/components/abstract-smart-component.js
@@ -8,13 +8,9 @@ export default class AbstractSmartComponent extends AbstractComponent {
   rerender() {
     const oldElement = this.getElement();
     const parent = oldElement.parentElement;
-
     this.removeElement();
-
     const newElement = this.getElement();
-
     parent.replaceChild(newElement, oldElement);
-
 
     this.recoveryListeners();
   }
